move forking tests to the local directory (bad merge)
