+++ test/local/forking/snapshot.js
@@ -1,6 +1,6 @@
 const assert = require("assert");
-const bootstrap = require("../helpers/contract/bootstrap");
-const intializeTestProvider = require("../helpers/web3/initializeTestProvider");
+const bootstrap = require("../../helpers/contract/bootstrap");
+const intializeTestProvider = require("../../helpers/web3/initializeTestProvider");
 
 /**
  * NOTE: Naming in these tests is a bit confusing. Here, the "main chain"
